consumer/requirements.txt
kafka-python
requests
boto3
opencv-python
pillow
python-dotenv
